+++ types/loginConfig/index.d.ts
@@ -10,4 +10,7 @@ export type LoginConfigStore = {
 	carbonioAdminUiDarkAppLogo?: string;
 	carbonioAdminUiTitle?: string;
 	carbonioAdminUiFavicon?: string;
+	carbonioLogoURL?: string;
+	carbonioWebUiPrimaryColor?: string;
+	carbonioWebUiDarkPrimaryColor?: string;
 };