+++ CHANGELOG.md
@@ -2,6 +2,44 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.9.12](https://github.com/zextras/carbonio-admin-ui/compare/v0.9.11...v0.9.12) (2023-04-27)
+
+
+### Bug Fixes
+
+* build issue ([be42044](https://github.com/zextras/carbonio-admin-ui/commit/be42044f7c70dc3689f799e40bdbd412bcacb4b6))
+
+### [0.9.11](https://github.com/zextras/carbonio-admin-ui/compare/v0.9.10...v0.9.11) (2023-04-12)
+
+
+### Bug Fixes
+
+* load local copy of font instead of external call ([#63](https://github.com/zextras/carbonio-admin-ui/issues/63)) ([74570ed](https://github.com/zextras/carbonio-admin-ui/commit/74570ed2bfcd010c24c5de257e11fe966b7e2a5d))
+
+### [0.9.10](https://github.com/zextras/carbonio-admin-ui/compare/v0.9.9...v0.9.10) (2023-03-30)
+
+
+### Bug Fixes
+
+* support put/delete api ([#59](https://github.com/zextras/carbonio-admin-ui/issues/59)) ([300d9cc](https://github.com/zextras/carbonio-admin-ui/commit/300d9cc5cc0f29caa190189571c0baf0235855a5))
+
+### [0.9.9](https://github.com/zextras/carbonio-admin-ui/compare/v0.9.8...v0.9.9) (2023-03-14)
+
+### [0.9.8](https://github.com/zextras/carbonio-admin-ui/compare/v0.9.7...v0.9.8) (2023-03-02)
+
+
+### Bug Fixes
+
+* json array is not passing in api ([7fdfce1](https://github.com/zextras/carbonio-admin-ui/commit/7fdfce161c19a513664c74d47973dc439f3b5328))
+
+### [0.9.7](https://github.com/zextras/carbonio-admin-ui/compare/v0.9.6...v0.9.7) (2023-02-02)
+
+
+### Bug Fixes
+
+* add logo url open in new page ([#45](https://github.com/zextras/carbonio-admin-ui/issues/45)) ([1990753](https://github.com/zextras/carbonio-admin-ui/commit/19907538c74391c8ebe47c43c02ace4a54fe2e8e))
+* splash screen logo of admin panel ([#46](https://github.com/zextras/carbonio-admin-ui/issues/46)) ([a88861c](https://github.com/zextras/carbonio-admin-ui/commit/a88861c9ac405d446492a1d497c1918359d5297e))
+
 ### [0.9.6](https://github.com/zextras/carbonio-admin-ui/compare/v0.9.5...v0.9.6) (2023-01-05)